requirements.txt
portia-sdk-python[google]>=0.6.2,<0.7
python-dotenv>=1.0.1,<2
pydantic>=2,<3